backend/pyproject.toml
[tool.poetry]
package-mode = false
name = "topicnet-backend"
version = "0.1.0"
description = "topicnet-backend is a FastAPI-based backend service designed to process unstructured text documents in PDF format and generate mind maps. Leveraging natural language processing techniques, the backend extracts key concepts and relationships from the text, visualizing them in an intuitive mind map format. This service provides APIs for uploading PDFs, analyzing their content, and generating structured visual representations, making it easier to understand and explore complex documents. Ideal for educational purposes, research, and information management."
authors = ["Dulina-Perera <dulina.perera@outlook.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aioboto3 = "^13.2.0"
alembic = "1.13.3"
asyncpg = "^0.29.0"
boto3 = "1.35.31"
fastapi = "0.115.0"
gensim = "4.3.3"
hdbscan = "0.8.38.post1"
huggingface-hub = "0.25.1"
openai = "1.47.1"
psycopg = "3.2.3"
PyMuPDF = "1.24.10"
pymupdf4llm = "0.0.17"
pytextrank = "3.3.0"
python-dotenv = "1.0.1"
python-multipart = "0.0.11"
scikit-learn = "1.5.2"
sentence-transformers = "3.1.1"
spacy = "3.7.5"
SQLAlchemy = "2.0.35"
torch = "2.4.1"
transformers = "4.45.2"
umap-learn = "0.5.6"
uvicorn = "0.31.0"

[tool.poetry.group.dev.dependencies]
icecream = "2.1.3"
thinc = "8.2.2"
tqdm = "4.66.5"
uv = "0.4.20"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
